// src/shared/services/api/models/courseSummaryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { RefTermDto } from './refTermDto';

export interface CourseSummaryDto {
  id?: string;
  /** @nullable */
  title?: string | null;
  category?: RefTermDto;
  difficulty_level?: RefTermDto;
  duration_minute?: number;
  overall_rating?: number;
  /** @nullable */
  thumbnail_img?: string | null;
  is_drafted?: boolean;
}