// src/shared/services/api/models/paginatedCoursesDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { CourseSummaryDto } from './courseSummaryDto';

export interface PaginatedCoursesDto {
  /** @nullable */
  data?: CourseSummaryDto[] | null;
  total?: number;
  page?: number;
  page_size?: number;
}